chore(): restructure exports and move types to top of source.
